Check if a module is already loaded before loading it from filter input. (#158)
Before this change, every time filter input is made then erlyberly will attempt to load that as a module. This helps to quickly trace modules that are not loaded yet in non-embedded releases. However, each input would reset the module tree making selection very difficult.

Now, erlyberly checks on the remote node that the module is not loaded before attempting to load it, so no call back for the module load is made when changing the filter for one module multiple times.