Record Metadata in Function Traces (#160)
Before this change, records and record fields would be highlighted for gen_server callbacks. Now records that are defined in the traced module (inc. included header files) will be highlighted in function traces.

When a trace call is made, the modules beam file is parsed and record AST is pulled out and stored in memory in the UI. When the term is displayed in the tree, if a matching record name within the traced module is found then the metadata labels are added in the tree.